Fixed a bug with deleting contacts that have no picture set

+++ contacts/views.py
@@ -92,7 +92,8 @@ def contact_delete(request,document_id):
 	contact = get_object_or_404(Contact,id=document_id)
 	company_id = contact.company_id
 	if user_can_access_company(request.user,company_id):
-		contact.picture.delete()
+		if (contact.picture):
+			contact.picture.delete()
 		contact.delete()
 		if request.is_ajax():
 			return HttpResponse(status=200)
